fix(migrations) avoid performing full updates
The previous intent was to rely on a schema's default field to insert
default values for new fields. However, a `full` update means we pass a
complete row to the DAO, and that row might contain older fields. Those
fields come from the row in the DB, but **disappeared** from the schema
definition. Hence, the DAO errors on the update and complains that
fields (such as `strip_request_path`) do not exist anymore, which is
correct, according to the current API schema.

Hence, we prefer not performing full updates, and only pass a subset of
entities' fields to `dao:update()`. Furthermore, we hard-code default
values, because using the schema (from `require`) would also result in
future circular-dependencies (if the filed is removed from schema) or
subject to wrong defaults being inserted if the `default` field in the
schema changes in the future.

Additionally, we avoid selecting fields that might not yet exist in
SELECT clauses (build from the schema).